+++ modelconverter/cli/utils.py
@@ -1,13 +1,23 @@
 import logging
 import shutil
+from contextlib import suppress
 from datetime import datetime
 from pathlib import Path
-from typing import Dict, List, Optional, Tuple
+from typing import Any, Dict, List, Literal, Optional, Tuple
+from uuid import UUID
 
 from luxonis_ml.nn_archive import is_nn_archive
 from luxonis_ml.nn_archive.config import Config as NNArchiveConfig
 from luxonis_ml.nn_archive.config_building_blocks import PreprocessingBlock
+from rich import print
+from rich.box import ROUNDED
+from rich.console import Console, Group
+from rich.markdown import Markdown
+from rich.panel import Panel
+from rich.pretty import Pretty
+from rich.table import Table
 
+from modelconverter.hub.hub_requests import Request
 from modelconverter.utils import (
     process_nn_archive,
     resolve_path,
@@ -109,3 +119,150 @@ def extract_preprocessing(
         inp.reverse_input_channels = False
 
     return cfg, preprocessing
+
+
+def print_hub_resource_info(
+    model: Dict[str, Any], keys: List[str], json: bool, **kwargs
+) -> Dict[str, Any]:
+    if json:
+        print(model)
+        return model
+
+    console = Console()
+
+    if model.get("description_short"):
+        description_short_panel = Panel(
+            f"[italic]{model['description_short']}[/italic]",
+            border_style="dim",
+            box=ROUNDED,
+        )
+    else:
+        description_short_panel = None
+
+    table = Table(show_header=False, box=None)
+    table.add_column(justify="right", style="bold")
+    table.add_column()
+    for key in keys:
+        if key in ["created", "updated", "last_version_added"]:
+            value = model.get(key, "N/A")
+
+            def format_date(date_str):
+                try:
+                    date_obj = datetime.strptime(
+                        date_str, "%Y-%m-%dT%H:%M:%S.%f"
+                    )
+                    return date_obj.strftime("%B %d, %Y %H:%M:%S")
+                except (ValueError, TypeError):
+                    return Pretty("N/A")
+
+            formatted_value = format_date(value)
+
+            table.add_row(f"{key.replace('_', ' ').title()}:", formatted_value)
+        elif key == "is_public":
+            if key not in model:
+                value = "N/A"
+            else:
+                value = "Public" if model[key] else "Private"
+            table.add_row("Visibility", Pretty(value))
+        elif key == "is_commercial":
+            # TODO: Is Usage the right term?
+            if key not in model:
+                value = "N/A"
+            else:
+                value = "Commercial" if model[key] else "Non-Commercial"
+            table.add_row("Usage:", Pretty(value))
+        elif key == "is_nn_archive":
+            table.add_row("NN Archive:", Pretty(model.get(key, False)))
+        else:
+            table.add_row(
+                f"{key.replace('_', ' ').title()}:",
+                Pretty(model.get(key, "N/A")),
+            )
+
+    info_panel = Panel(table, border_style="cyan", box=ROUNDED, **kwargs)
+
+    if model.get("description"):
+        description_panel = Panel(
+            Markdown(model["description"]),
+            title="Description",
+            border_style="green",
+            box=ROUNDED,
+        )
+    else:
+        description_panel = None
+
+    nested_panels = []
+    if description_short_panel:
+        nested_panels.append(description_short_panel)
+    nested_panels.append(info_panel)
+    if description_panel:
+        nested_panels.append(description_panel)
+
+    content = Group(*nested_panels)
+
+    main_panel = Panel(
+        content,
+        title=f"[bold magenta]{model.get('name', 'N/A')}[/bold magenta]",
+        width=74,
+        border_style="magenta",
+        box=ROUNDED,
+    )
+
+    console.print(main_panel)
+    console.rule()
+    return model
+
+
+def hub_ls(endpoint: str, keys: List[str], **kwargs) -> None:
+    data = Request.get(f"/api/v1/{endpoint}/", params=kwargs).json()
+    print(kwargs)
+    table = Table(row_styles=["yellow", "cyan"], box=ROUNDED)
+    for key in keys:
+        table.add_column(key, header_style="magenta i")
+
+    for model in data:
+        renderables = []
+        for key in keys:
+            value = model.get(key, "N/A")
+            if isinstance(value, list):
+                value = ", ".join(value)
+            renderables.append(value)
+        table.add_row(*renderables)
+
+    console = Console()
+    console.print(table)
+
+
+def is_valid_uuid(uuid_string: str) -> bool:
+    try:
+        UUID(uuid_string)
+        return True
+    except Exception:
+        return False
+
+
+def slug_to_id(
+    slug: str, endpoint: Literal["models", "modelVersions", "modelInstances"]
+) -> str:
+    for is_public in [True, False]:
+        with suppress(Exception):
+            params = {
+                "is_public": is_public,
+                "slug": slug,
+            }
+            data = Request.get(f"/api/v1/{endpoint}/", params=params).json()
+            if data:
+                return data[0]["id"]
+    raise ValueError(f"Model with slug '{slug}' not found.")
+
+
+def request_info(
+    identifier: str,
+    endpoint: Literal["models", "modelVersions", "modelInstances"],
+) -> Dict[str, Any]:
+    if is_valid_uuid(identifier):
+        resource_id = identifier
+    else:
+        resource_id = slug_to_id(identifier, endpoint)
+
+    return Request.get(f"/api/v1/{endpoint}/{resource_id}/").json()
